+++ Sponge/src/net/tnemc/sponge/command/MoneyCommand.java
@@ -19,6 +19,7 @@ package net.tnemc.sponge.command;
  */
 
 import net.tnemc.core.account.Account;
+import net.tnemc.core.command.parameters.PercentBigDecimal;
 import net.tnemc.core.currency.Currency;
 import net.tnemc.sponge.impl.SpongeCMDSource;
 import revxrsal.commands.annotation.Command;
@@ -62,7 +63,7 @@ public class MoneyCommand {
   @Usage("#{Money.Convert.Arguments}")
   @Description("#{Money.Convert.Description}")
   @CommandPermission("tne.money.convert")
-  public void onConvert(SpongeCommandActor sender, BigDecimal amount, Currency currency, Currency fromCurrency) {
+  public void onConvert(SpongeCommandActor sender, PercentBigDecimal amount, Currency currency, Currency fromCurrency) {
     net.tnemc.core.command.MoneyCommand.onConvert(new SpongeCMDSource(sender), amount, currency, fromCurrency);
   }
 
@@ -70,7 +71,7 @@ public class MoneyCommand {
   @Usage("Money.Deposit.Arguments")
   @Description("Money.Deposit.Description")
   @CommandPermission("tne.money.deposit")
-  public void onDeposit(SpongeCommandActor sender, BigDecimal amount, @Default("") Currency currency, @Default("world-113") String region) {
+  public void onDeposit(SpongeCommandActor sender, PercentBigDecimal amount, @Default("") Currency currency, @Default("world-113") String region) {
     net.tnemc.core.command.MoneyCommand.onDeposit(new SpongeCMDSource(sender), amount, currency, region);
   }
 
@@ -78,7 +79,7 @@ public class MoneyCommand {
   @Usage("#{Money.Give.Arguments}")
   @Description("#{Money.Give.Description}")
   @CommandPermission("tne.money.give")
-  public void onGive(SpongeCommandActor sender, Account player, BigDecimal amount, @Default("world-113") String region, @Default("") Currency currency) {
+  public void onGive(SpongeCommandActor sender, Account player, PercentBigDecimal amount, @Default("world-113") String region, @Default("") Currency currency) {
     net.tnemc.core.command.MoneyCommand.onGive(new SpongeCMDSource(sender), player, amount, region, currency);
   }
 
@@ -86,7 +87,7 @@ public class MoneyCommand {
   @Usage("#{Money.Note.Arguments}")
   @Description("#{Money.Note.Description}")
   @CommandPermission("tne.money.note")
-  public void onNote(SpongeCommandActor sender, BigDecimal amount, @Default("") Currency currency) {
+  public void onNote(SpongeCommandActor sender, PercentBigDecimal amount, @Default("") Currency currency) {
     net.tnemc.core.command.MoneyCommand.onNote(new SpongeCMDSource(sender), amount, currency);
   }
 
@@ -102,7 +103,7 @@ public class MoneyCommand {
   @Usage("#{Money.Pay.Arguments}")
   @Description("#{Money.Pay.Description}")
   @CommandPermission("tne.money.pay")
-  public void onPay(SpongeCommandActor sender, Account player, BigDecimal amount, @Default("") Currency currency, @Default("") String from) {
+  public void onPay(SpongeCommandActor sender, Account player, PercentBigDecimal amount, @Default("") Currency currency, @Default("") String from) {
     net.tnemc.core.command.MoneyCommand.onPay(new SpongeCMDSource(sender), player, amount, currency, from);
   }
 
@@ -134,7 +135,7 @@ public class MoneyCommand {
   @Usage("#{Money.Take.Arguments}")
   @Description("#{Money.Take.Description}")
   @CommandPermission("tne.money.take")
-  public void onTake(SpongeCommandActor sender, Account player, BigDecimal amount, @Default("world-113") String region, @Default("") Currency currency) {
+  public void onTake(SpongeCommandActor sender, Account player, PercentBigDecimal amount, @Default("world-113") String region, @Default("") Currency currency) {
     net.tnemc.core.command.MoneyCommand.onTake(new SpongeCMDSource(sender), player, amount, region, currency);
   }
 
@@ -150,7 +151,7 @@ public class MoneyCommand {
   @Usage("Money.Withdraw.Arguments")
   @Description("Money.Withdraw.Description")
   @CommandPermission("tne.money.withdraw")
-  public void onWithdraw(SpongeCommandActor sender, BigDecimal amount, @Default("") Currency currency, @Default("world-113") String region) {
+  public void onWithdraw(SpongeCommandActor sender, PercentBigDecimal amount, @Default("") Currency currency, @Default("world-113") String region) {
     net.tnemc.core.command.MoneyCommand.onWithdraw(new SpongeCMDSource(sender), amount, currency, region);
   }
 }