+++ changelog.md
@@ -7,3 +7,4 @@
 - Added Redis Support. This will allow data syncing without the need for bungee
 - Added a currency template module(requested by stoffeh)
   - This contains some example pre-configured commonly used currency setups for TNE.
+- Added $currency, $player variables for /money give/take messages in messages.yml